- Decoupled the MoreViewController into 3 different files
