+++ ownCloud/Client/Actions/MoreViewController.swift
@@ -19,89 +19,6 @@
 import UIKit
 import ownCloudSDK
 
-class MoreViewHeader: UIView {
-	private var iconView: UIImageView
-	private var titleLabel: UILabel
-	private var detailLabel: UILabel
-
-	var item: OCItem
-	weak var core: OCCore?
-
-	init(for item: OCItem, with core: OCCore) {
-		self.item = item
-		self.core = core
-
-		iconView = UIImageView()
-		titleLabel = UILabel()
-		detailLabel = UILabel()
-		super.init(frame: .zero)
-
-		render()
-	}
-
-	private func render() {
-		titleLabel.translatesAutoresizingMaskIntoConstraints = false
-		detailLabel.translatesAutoresizingMaskIntoConstraints = false
-		iconView.translatesAutoresizingMaskIntoConstraints = false
-		iconView.contentMode = .scaleAspectFit
-
-		titleLabel.font = UIFont.systemFont(ofSize: 17, weight: UIFont.Weight.semibold)
-		detailLabel.font = UIFont.systemFont(ofSize: 14)
-
-		detailLabel.textColor = UIColor.gray
-
-		self.addSubview(titleLabel)
-		self.addSubview(detailLabel)
-		self.addSubview(iconView)
-
-		iconView.leftAnchor.constraint(equalTo: self.leftAnchor, constant: 20).isActive = true
-		iconView.rightAnchor.constraint(equalTo: titleLabel.leftAnchor, constant: -15).isActive = true
-		iconView.rightAnchor.constraint(equalTo: detailLabel.leftAnchor, constant: -15).isActive = true
-
-		titleLabel.rightAnchor.constraint(equalTo:  self.rightAnchor, constant: -20).isActive = true
-		detailLabel.rightAnchor.constraint(equalTo: self.rightAnchor, constant: -20).isActive = true
-
-		iconView.widthAnchor.constraint(equalToConstant: 60).isActive = true
-		iconView.centerYAnchor.constraint(equalTo: self.centerYAnchor).isActive = true
-
-		titleLabel.topAnchor.constraint(equalTo: self.topAnchor, constant: 20).isActive = true
-		titleLabel.bottomAnchor.constraint(equalTo: detailLabel.topAnchor, constant: -5).isActive = true
-		detailLabel.bottomAnchor.constraint(equalTo: self.bottomAnchor, constant: -20).isActive = true
-
-		iconView.setContentHuggingPriority(UILayoutPriority.required, for: UILayoutConstraintAxis.vertical)
-		titleLabel.setContentCompressionResistancePriority(UILayoutPriority.defaultHigh, for: UILayoutConstraintAxis.vertical)
-		detailLabel.setContentCompressionResistancePriority(UILayoutPriority.defaultHigh, for: UILayoutConstraintAxis.vertical)
-
-		titleLabel.attributedText = NSAttributedString(string: item.name, attributes: [NSAttributedStringKey.font: UIFont.systemFont(ofSize: 17, weight: .semibold)])
-		detailLabel.attributedText =  NSAttributedString(string: item.name, attributes: [NSAttributedStringKey.font: UIFont.systemFont(ofSize: 17, weight: .semibold)])
-
-		self.iconView.image = item.icon(fitInSize: CGSize(width: 40, height: 40))
-
-		if item.thumbnailAvailability != .none {
-			let displayThumbnail = { (thumbnail: OCItemThumbnail?) in
-				_ = thumbnail?.requestImage(for: CGSize(width: 60, height: 60), scale: 0, withCompletionHandler: { (thumbnail, error, _, image) in
-					if error == nil,
-						image != nil,
-						self.item.itemVersionIdentifier == thumbnail?.itemVersionIdentifier {
-						OnMainThread {
-							self.iconView.image = image
-						}
-					}
-				})
-			}
-
-			_ = core?.retrieveThumbnail(for: item, maximumSize: CGSize(width: 150, height: 150), scale: 0, retrieveHandler: { (error, _, _, thumbnail, _, progress) in
-				displayThumbnail(thumbnail)
-			})
-		}
-		titleLabel.numberOfLines = 0
-	}
-
-	required init?(coder aDecoder: NSCoder) {
-		fatalError("init(coder:) has not been implemented")
-	}
-}
-
 class MoreViewController: UIViewController {
 
 	private var item: OCItem
@@ -179,57 +96,5 @@ class MoreViewController: UIViewController {
 extension MoreViewController: Themeable {
 	func applyThemeCollection(theme: Theme, collection: ThemeCollection, event: ThemeEvent) {
 		self.headerView.backgroundColor = Theme.shared.activeCollection.tableBackgroundColor
-	}
-}
-
-class MoreStaticTableViewController: StaticTableViewController {
-
-	override func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-		if let title = sections[section].headerAttributedTitle {
-			let containerView = UIView()
-			let label = UILabel()
-			label.translatesAutoresizingMaskIntoConstraints = false
-			containerView.addSubview(label)
-			NSLayoutConstraint.activate([
-				label.leftAnchor.constraint(equalTo: containerView.leftAnchor, constant: 32),
-				label.topAnchor.constraint(equalTo: containerView.topAnchor),
-				label.bottomAnchor.constraint(equalTo: containerView.bottomAnchor),
-				label.rightAnchor.constraint(equalTo: containerView.rightAnchor, constant: -32)
-				])
-
-			label.attributedText = title
-			return containerView
-		}
-
-		return nil
-	}
-
-	override func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-		if indexPath.section != 0 {
-			return 56
-		}
-
-		return UITableViewAutomaticDimension
-	}
-
-	override func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-		if sections[section].headerAttributedTitle != nil || sections[section].headerTitle != nil {
-			return 56
-		}
-
-		return 0
-	}
-
-	override func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
-		if sections[section].footerAttributedTitle != nil || sections[section].footerTitle != nil {
-			return 56
-		}
-
-		return 0
-	}
-
-	override func applyThemeCollection(theme: Theme, collection: ThemeCollection, event: ThemeEvent) {
-		super.applyThemeCollection(theme: theme, collection: collection, event: event)
-		self.tableView.separatorColor = self.tableView.backgroundColor
 	}
 }
